Removed the multiple login pages and made it one
