+++ src/app/get_started/page.tsx
@@ -2,21 +2,22 @@ import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
 import Link from "next/link";
 
-export default function GetStarted() {
+export default function GetStarted() 
+{
   return (
-    <div className="container mx-auto flex flex-col items-center justify-center min-h-screen px-3">
+    <div className="flex flex-col items-center justify-center min-h-screen px-3">
       <h1 className="text-3xl font-bold mb-16 text-center">Get Started</h1>
 
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-8 w-full max-w-4xl">
+      <div className="w-full max-w-4xl">
         <Card className="bg-gray-200">
           <CardContent className="flex flex-col items-center justify-center p-8">
-            <h2 className="text-xl font-bold mb-6">Farmer</h2>
+            {/* <h2 className="text-xl font-bold mb-6">Farmer</h2> */}
 
             <Button
               asChild
               className="w-full max-w-xs bg-blue-500 hover:bg-blue-600 text-white mb-4"
             >
-              <Link href="/signup/farmer">Create Account</Link>
+              <Link href="/signup">Create Account</Link>
             </Button>
 
             <Button
@@ -24,12 +25,12 @@ export default function GetStarted() {
               variant="outline"
               className="w-full max-w-xs bg-blue-500 hover:bg-blue-600 text-white"
             >
-              <Link href="/login/farmer">Login Instead</Link>
+              <Link href="/login">Login Instead</Link>
             </Button>
           </CardContent>
         </Card>
 
-        <Card className="bg-gray-200">
+        {/* <Card className="bg-gray-200">
           <CardContent className="flex flex-col items-center justify-center p-8">
             <h2 className="text-xl font-bold mb-6">Administrator</h2>
 
@@ -48,7 +49,7 @@ export default function GetStarted() {
               <Link href="/login/admin">Login Instead</Link>
             </Button>
           </CardContent>
-        </Card>
+        </Card> */}
       </div>
     </div>
   );
